move loading of dashboard providers to listeners again
